# packages/sidflow-web/openapi.yaml
openapi: 3.0.3
info:
  title: SIDFlow Web API
  description: |
    Local web interface for SIDFlow.

    The API primarily orchestrates server-side helpers and shared package logic (it does not shell out to the interactive CLIs).
    See `packages/sidflow-web/app/api/**/route.ts` for the concrete implementation.
  version: 0.3.10
  license:
    name: GPL-2.0-only
    url: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
servers:
  - url: http://localhost:3000
    description: Default local development server (override per environment)
security: []
...
paths:
...
  /api/classify:
    post:
      tags:
        - classification
      summary: Classify SID files
      description: |
        Trigger classification on a directory of SID files.

        The default web flow sends `async: true`, which queues a durable background job and returns `202 Accepted`.
        Synchronous execution remains available for callers that omit `async`.
      operationId: classifySids
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyRequest"
            example:
              path: /path/to/sid/directory
              async: true
      responses:
        "200":
          description: Classification completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "202":
          description: Classification job queued successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ClassifyQueuedResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/CommandError"
        "500":
          $ref: "#/components/responses/CommandError"
...
  /api/train:
    post:
      tags:
        - training
      summary: Train ML model
      description: Queue a durable training job for TensorFlow.js model artifacts.
      operationId: trainModel
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainRequest"
            examples:
              default:
                summary: Default training
                value: {}
              custom:
                summary: Custom training parameters
                value:
                  epochs: 20
                  batchSize: 32
                  learningRate: 0.0005
                  evaluate: true
                  force: false
      responses:
        "202":
          description: Training job queued successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TrainQueuedResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/CommandError"
        "500":
          $ref: "#/components/responses/CommandError"
components:
  schemas:
...
    ClassifyRequest:
      type: object
      properties:
        path:
          type: string
          description: Path to directory containing SID files to classify
          minLength: 1
          example: /home/user/hvsc/MUSICIANS/H
        forceRebuild:
          type: boolean
          description: Force rerendering even when cached WAV data exists.
        skipAlreadyClassified:
          type: boolean
          description: Skip entries already present in the auto-tag output.
        deleteWavAfterClassification:
          type: boolean
          description: Delete rendered WAV files after classification to reduce disk use.
        async:
          type: boolean
          description: Queue the classification as a durable background job.
          default: false
...
    ClassifyProgress:
      type: object
      additionalProperties: true
    ClassifyQueuedResponse:
      type: object
      required:
        - started
        - jobId
        - progress
      properties:
        started:
          type: boolean
          enum:
            - true
        jobId:
          type: string
        progress:
          $ref: "#/components/schemas/ClassifyProgress"
    TrainQueuedResponse:
      type: object
      required:
        - jobId
        - output
        - logs
      properties:
        jobId:
          type: string
        output:
          type: string
        logs:
          type: string
...
    TrainRequest:
      type: object
      properties:
        configPath:
          type: string
          description: Path to custom .sidflow.json configuration file
          example: /path/to/config.json
        epochs:
          type: integer
          description: Number of training epochs
          minimum: 1
          example: 10
        batchSize:
          type: integer
          description: Training batch size
          minimum: 1
          example: 16
        learningRate:
          type: number
          description: Learning rate for training
          minimum: 0
          example: 0.001
        evaluate:
          type: boolean
          description: Whether to evaluate on test set
          default: true
          example: true
        force:
          type: boolean
          description: Force complete retraining from scratch
          default: false
          example: false
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum:
            - true
          example: true
        data:
          type: object
          properties:
            output:
              type: string
              description: Standard output from CLI command
              example: "Playlist generated successfully\nPlaying: Commando.sid\n"
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
          example: false
        error:
          type: string
          description: Error message
          example: Validation error
        details:
          type: string
          description: Additional error details
          example: "sid_path: SID path is required"
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: Validation error
            details: "ratings.e: Expected integer, received string"
    CommandError:
      description: CLI command execution failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: Playback command failed
            details: "Error: SID file not found"